scripts/requirements.txt
Faker==22.2.0
psycopg2-binary==2.9.9
pytest==7.4.4
python-dotenv==1.0.1
alembic==1.13.1